Extend installer to Debian-based distros

+++ installers/servicestart.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
-# When wireless client AP mode is enabled, this script handles starting
+# When wireless client AP or Bridge mode is enabled, this script handles starting
 # up network services in a specific order and timing to avoid race conditions.
 
 PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
-NAME=raspap
+NAME=raspapd
 DESC="Service control for RaspAP"
 CONFIGFILE="/etc/raspap/hostapd.ini"
-DAEMONPATH="/lib/systemd/system/raspap.service"
+DAEMONPATH="/lib/systemd/system/raspapd.service"
 OPENVPNENABLED=$(pidof openvpn | wc -l)
 
 positional=()
@@ -22,21 +22,33 @@ case $key in
     ;;
     -s|--seconds)
     seconds="$2"
-    shift # past argument
-    shift # past value
+    shift
+    shift
+    ;;
+    -a|--action)
+    action="$2"
+    shift
+    shift
     ;;
 esac
 done
 set -- "${positional[@]}"
 
 echo "Stopping network services..."
-systemctl stop openvpn-client@client
+if [ $OPENVPNENABLED -eq 1 ]; then
+    systemctl stop openvpn-client@client
+fi
 systemctl stop systemd-networkd
 systemctl stop hostapd.service
 systemctl stop dnsmasq.service
 systemctl stop dhcpcd.service
 
-if [ -f "$DAEMONPATH" ]; then
+if [ "${action}" = "stop" ]; then
+    echo "Services stopped. Exiting."
+    exit 0
+fi
+
+if [ -f "$DAEMONPATH" ] && [ ! -z "$interface" ]; then
     echo "Changing RaspAP Daemon --interface to $interface"
     sed -i "s/\(--interface \)[[:alnum:]]*/\1$interface/" "$DAEMONPATH"
 fi
@@ -49,6 +61,9 @@ if [ -r "$CONFIGFILE" ]; then
 
     if [ "${config[BridgedEnable]}" = 1 ]; then
         if [ "${interface}" = "br0" ]; then
+            echo "Stopping systemd-networkd"
+            systemctl stop systemd-networkd
+
             echo "Restarting eth0 interface..."
             ip link set down eth0
             ip link set up eth0
@@ -64,14 +79,21 @@ if [ -r "$CONFIGFILE" ]; then
         echo "Disabling systemd-networkd"
         systemctl disable systemd-networkd
 
-        echo "Removing br0 interface..."
-        ip link set down br0
-        ip link del dev br0
+        ip link ls up | grep -q 'br0' &> /dev/null
+        if [ $? == 0 ]; then
+            echo "Removing br0 interface..."
+            ip link set down br0
+            ip link del dev br0
+        fi
 
         if [ "${config[WifiAPEnable]}" = 1 ]; then
             if [ "${interface}" = "uap0" ]; then
-                echo "Removing uap0 interface..."
-                iw dev uap0 del
+
+                ip link ls up | grep -q 'uap0' &> /dev/null
+                if [ $? == 0 ]; then
+                    echo "Removing uap0 interface..."
+                    iw dev uap0 del
+                fi
 
                 echo "Adding uap0 interface to ${config[WifiManaged]}"
                 iw dev ${config[WifiManaged]} interface add uap0 type __ap
